Make readdir return DT_LNK for symlinks.
This commit also makes readdir fill dirent::d_ino with inode found in Unix EAs (if present) rather than randomly generate it on its own. The benefit of this is that now inode returned by readdir is in-line with what stat returns. It was a long todo as well.

See #9 for more details.