Add LIBC066 forwarder.
This forwarder is needed since we are going to change the DLL name in #17. Besides forwarding LIBC stuff to the actual DLL, this forwarder also includes and exports the standard C++ library from GCC 3.
This is necessary because the original LIBC062..066 DLLs export C++ due to the lack of a separate GCC 3 libstdc++ DLL.

Needed for #4.