Properly define kHeapDbgException as weak symbol.
GCC4 assumes that strong functions' addresses are never NULL so comparing them to NULL always fails.
kHeapDbgException was declared weak beyond GCC control using direct assembly so it thought it's a strong symbol.
As a result, it would generate a call to it even when it was NULL at runtime causing SIGILL.

GCC3 would allow functions to be NULL so it worked as designed.

The fix is to use #pragma weak to do it the GCC way.

Closes #26.