Use -Wno-strict-aliasing everywhere.
GCC 4 barfs with
```
warning: dereferencing type-punned pointer will break strict-aliasing rules
```
when converting data addresses to pointers of incompatible types.

This, however, doesn't make any sense in case of LIBC as it is (so far) only intended to be built on 32-bit OS/2 where char is always 8 bit, all system structs are pack(1)'ed
and such pointer conversions are generally safe (and used a lot since this was how things were done back then).

Needed for #21.