Make _fnlwr/_fnlwr2 work for path rewrites (/@unixroot and such).
It would incorrectly detect FS type for such paths if the current
drive didn't match the original path drive (e.g. was a CDROM). This
would in turn lead to really fancy bugs like python not being able
to find UserDict.py and other mixed case modules when started from
a directory of a CDROM drive.

This commit also removes unconditional lowering of the drive letter
which happened in readdir (which uses _fnlwr) even on FSes supporting
mixed case.

Fixes #39.