Use gcc instead of makedep to generate dependency targets for makefiles.
This can be done from project root with `make dep` (`make cleandep` deletes
all generated dependencies).

Note that this still works the old makepdep way. I.e. you need to run
`make dep` after a clean checkout or when new include statements are added
or removed from sources. This generates dependencies for all files at
once and doesn't need to be repeated each time the source or header is
changed (unless a new include is added or removed, of course).

Closes #56.