logstrict: Create log files in system log directory by default.
The system log directory for LIBC is either LOGFILES\app or UNIXROOT\var\log\app (if LOGFILES is not set) or root of the boot drive as a fallback (if creating the "app" subdirectory fails).

The log filename is also changed to avoid filename conflicts where many different processes write logs to the same directory over a long period of time.

There is also a new variable LIBC_LOGGING_OUTPUT which can be set to stdout/stderr to redirect logging to the console, or to curdir to force log file creation in the current directory (old behavior).

Also, the log file header now prints more information (e.g. enabled log groups, all process arguments instead of just two).

This commit also introduces a number of additions to the LIBC logging API (including __libc_LogInitEx) to give the user more control over it, refer to API docs in headers for details.

Needed for #101.