logstrict: Print PID:TID and ORIGIN:GROUP when logging to console.
This is instead of just TID and GROUP to distinguish between logs from different processes and libraries/apps.

This commit also adds new API exports: __libc_LogVSNPrintf and __libc_LogSNPrintf that allow to use special format specifiers to do the same thing from user loggers. Note that these function are not marked printf-like to avoid warnings due to non-standard format specifiers so they should be used with care and only to prepare log entries.

Needed for #101.