Move EXCEPTQ .TRP reports generated by LIBC to system log directory.
Note that when moving, logs get a different name which is similar to LIBC log files (prepended by a timestamp to avoid name clashes) and a .txt extension.

If LIBC_LOGGING_OUTPUT is set (to redirect logs to curdir/stdio/stderr), moving .TRP files is disabled.

This is a temporary solution for item 3 from #99.

Also related to #101.