logstrict: Create reports/logs in BOOTDRIVE\var\log\app if no LOGFILES or UNIXROOT set.
It is not uncommon when a child exe is started with a clear enviornment. In this case,
LOGFILES and UNIXROOT will be reset even if they are set globally and crash reports
and logs will go to the root directory of the boot drive in this case polluting it.
Now they will go \var\log\app if \var\log exists (which is a typical setup anyway).

See #115 for a real life case.