startup: flush streams explicitly in _CRT_term()
This is essentially a move of the fix from LIBCx (see https://github.com/bitwiseworks/libcx/issues/31)
into LIBCn where it belongs.

This is a slightly modified patch by KO Myung-Hun <komh@chollian.net> (@komh).

Closes #139.

This is the original patch description that explains the problem in detail:

This fixes that buffered data of streams associatd with a socket are
lost.

This problem occurs because _CRT_term() calls weak terminators only if
its reference count is zero. However, all kLIBC programs are linked to
libc DLL dynamically. As a result, the count is not zero even if a
program ends, because libc DLL is not terminated, yet. That is,
_CRT_term() is not called yet in its _DLL_InitTerm(). Calling order is
like:

    1. _CRT_init() in _DLL_InitTerm() of libc
    2. _CRT_init() in crt0.s of a program
    3. _CRT_term() in exit() of a program
    4. Exit list routines registered by DosExitList()
    5. _CRT_term() in _DLL_InitTerm() of libc

After all, a program closes all the opened sockets in exit list of TCPIP
on step 4, but buffered data of streams are not flushed yet, because
_CRT_term() does not call __crtexit1__ set vector at this time. Flushing
is done when libc DLL is terminated on step 5. However, sockets are not
valid any more. As a result, buffered data of streams are lost.

And both exit() and abort() should flush streams, and they both calls
_CRT_term(). Therefore, flushing in _CRT_term() is proper. In addition,
it is not harmful even if calling flushall() too many.