tcpip: Call so_cancel when closing OS/2 socket.
This helps unblocking threads waiting on sockets with EINTR or EBADF
when these sockets are being in use by other processes and therefore
are not actually closed.

This commit also disables assertions on failed removesocketfromlist
calls in LOG builds which seem to be due to OS/2 TCP/IP DLL races.

See #150 for details.